Make views a subpackage
