+++ dimod/views/bqm.py
@@ -12,68 +12,11 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 #
-# ================================================================================================
+# =============================================================================
 try:
     import collections.abc as abc
 except ImportError:
     import collections as abc
-
-from collections import OrderedDict
-
-from six.moves import zip
-
-
-class IndexView(abc.Mapping):
-    __slots__ = '_variables', '_data'
-
-    def __init__(self, variables, data):
-        self._variables = variables
-        self._data = data
-
-    def __getitem__(self, v):
-        return self._data[self._variables.index(v)]
-
-    def __iter__(self):
-        return iter(self._variables)
-
-    def __len__(self):
-        return len(self._variables)
-
-    def __repr__(self):
-        return '{}({}, {})'.format(self.__class__.__name__, self._variables, self._data)
-
-    def __str__(self):
-        return str(dict(self))
-
-    def values(self):
-        return IndexValuesView(self)
-
-    def items(self):
-        return IndexItemsView(self)
-
-
-class SampleView(IndexView):
-    """View each row of the samples record as if it was a dict."""
-    def __repr__(self):
-        return str(self)
-
-
-class IndexItemsView(abc.ItemsView):
-    """Faster read access to the numpy array"""
-    __slots__ = ()
-
-    def __iter__(self):
-        # Inherited __init__ puts the Mapping into self._mapping
-        return zip(self._mapping._variables, self._mapping._data.flat)
-
-
-class IndexValuesView(abc.ValuesView):
-    """Faster read access to the numpy array"""
-    __slots__ = ()
-
-    def __iter__(self):
-        # Inherited __init__ puts the Mapping into self._mapping
-        return iter(self._mapping._data.flat)
 
 
 class BQMView(object):
